(Test + Networking) restructure middleware tests
